dhanush

+++ src/main/java/com/vts/pfms/committee/controller/ActionController.java
@@ -348,6 +348,7 @@ public class ActionController {
 		String UserId = (String) ses.getAttribute("Username");
 		logger.info(new Date() +"Inside ActionSubLaunch.htm "+UserId);		
 		try {		 
+			
 			String AssignerName=req.getParameter("Assigner");
 			req.setAttribute("Assignee", service.AssigneeData(req.getParameter("ActionMainId") ,req.getParameter("ActionAssignid")).get(0));
 			req.setAttribute("SubList", service.SubList(req.getParameter("ActionAssignid")));
@@ -376,8 +377,7 @@ public class ActionController {
 		try {
 			
 			
-			redir.addFlashAttribute("ActionMainId", req.getParameter("ActionMainId"));
-			redir.addFlashAttribute("ActionAssignId", req.getParameter("ActionAssignId"));
+			
 			ActionSubDto subDto=new ActionSubDto();
 			subDto.setLabCode(labCode);
 			subDto.setFileName(req.getParameter("FileName"));
@@ -396,7 +396,8 @@ public class ActionController {
 				redir.addAttribute("resultfail", "Action Update Unsuccessful");
 			}
 
-		
+			redir.addFlashAttribute("ActionMainId", req.getParameter("ActionMainId"));
+			redir.addFlashAttribute("ActionAssignId", req.getParameter("ActionAssignId"));
 		
 		}
 		catch (Exception e) {
@@ -432,12 +433,11 @@ public class ActionController {
 		     
 		     
 		     req.setAttribute("Assignee", data);
-		     req.setAttribute("SubList", service.SubList(data[19].toString()));
+		     req.setAttribute("SubList", service.SubList(data[18].toString()));
 		     req.setAttribute("AssignerName", AssignerName);
-			 req.setAttribute("LinkList", service.SubList(""));
-			 req.setAttribute("actiono",data[10].toString() );
+			 req.setAttribute("actiono",data[9].toString());
 			 req.setAttribute("filesize",file_size);
-			 List<Object[]> AssigneeDetails=service.AssigneeDetails(data[19].toString());
+			 List<Object[]> AssigneeDetails=service.AssigneeDetails(data[18].toString());
 			 
 			 if(AssigneeDetails.size()>0) 
 			 {
@@ -524,22 +524,23 @@ public class ActionController {
 		String UserId = (String) ses.getAttribute("Username");
 		logger.info(new Date() +"Inside ActionSubDelete.htm "+UserId);		
 		try { 
-			ActionAttachment attach=service.ActionAttachmentDownload(req.getParameter("ActionAttachid" ));	
-			int count = service.ActionSubDelete(req.getParameter("ActionSubId"), UserId);
-				
-			File my_file=null;
-				       
-			my_file = new File(uploadpath+attach.getAttachFilePath()+File.separator+attach.getAttachName()); 
+			ActionAttachment attach = null;
+			if(req.getParameter("ActionAttachid" )!=null) {
+				attach=service.ActionAttachmentDownload(req.getParameter("ActionAttachid" ));	
+				File my_file=null;
 			       
-			if(my_file.exists()) {
-				my_file.delete();
+				my_file = new File(uploadpath+attach.getAttachFilePath()+File.separator+attach.getAttachName()); 
+				       
+				if(my_file.exists()) {
+					my_file.delete();
+				}
 			}
-				
+			int count = service.ActionSubDelete(req.getParameter("ActionSubId"), UserId);	
 			if (count > 0) {
 				redir.addAttribute("result", "Action Sub Deleted Successfully");
 			} else {
 				redir.addAttribute("resultfail", "Action Sub Delete Unsuccessful");
-					}
+			}
 			redir.addFlashAttribute("ActionMainId", req.getParameter("ActionMainId"));
 			redir.addFlashAttribute("ActionAssignId",req.getParameter("ActionAssignId"));
 			return "redirect:/ActionSubLaunchRedirect.htm";
@@ -592,11 +593,11 @@ public class ActionController {
 						if(type.equalsIgnoreCase("A")) {
 							req.setAttribute("ForwardList", service.ForwardList(EmpId));
 						}else if (type.equalsIgnoreCase("NB")) {
-							req.setAttribute("ForwardList", service.ForwardList(EmpId).stream().filter(flag-> flag[7].toString().equalsIgnoreCase("N") || flag[7].toString().equalsIgnoreCase("B")).collect(Collectors.toList()));
+							req.setAttribute("ForwardList", service.ForwardList(EmpId).stream().filter(flag-> flag[6].toString().equalsIgnoreCase("A") || flag[7].toString().equalsIgnoreCase("B")).collect(Collectors.toList()));
 						}
 						req.setAttribute("type", type);
 					}else{
-						req.setAttribute("ForwardList", service.ForwardList(EmpId).stream().filter(flag-> flag[7].toString().equalsIgnoreCase("F")).collect(Collectors.toList()));
+						req.setAttribute("ForwardList", service.ForwardList(EmpId).stream().filter(flag-> flag[6].toString().equalsIgnoreCase("F")).collect(Collectors.toList()));
 						req.setAttribute("type", "F");
 					}
 
@@ -619,7 +620,7 @@ public class ActionController {
 			req.setAttribute("Assignee", service.AssigneeData(req.getParameter("ActionMainId"),req.getParameter("ActionAssignId")).get(0));
 			req.setAttribute("SubList", service.SubList(req.getParameter("ActionAssignId")));
 			req.setAttribute("AssigneeName", AssigneeName);
-			req.setAttribute("LinkList", service.SubList(req.getParameter("ActionLinkId")));
+			//req.setAttribute("LinkList", service.SubList(req.getParameter("ActionLinkId")));
 			req.setAttribute("actionslist", service.ActionSubLevelsList(req.getParameter("ActionAssignId")));
 			}
 			catch (Exception e) {
@@ -1258,24 +1259,17 @@ public class ActionController {
 					} 	
 		 
 					@RequestMapping(value = "ActionSelfList.htm", method = {RequestMethod.GET,RequestMethod.POST})
-					public String ActionSelf(HttpServletRequest req, HttpSession ses, RedirectAttributes redir)throws Exception {
-						
+					public String ActionSelf(HttpServletRequest req, HttpSession ses, RedirectAttributes redir)throws Exception 
+					{	
 						String UserId = (String) ses.getAttribute("Username");
 						logger.info(new Date() +"Inside ActionSelfList.htm "+UserId);		
 						try {
-						String EmpId = ((Long) ses.getAttribute("EmpId")).toString();
-						
-						
-
-						req.setAttribute("AssignedList", service.ActionSelfList(EmpId));
-						
-						}
-						catch (Exception e) {
+							String EmpId = ((Long) ses.getAttribute("EmpId")).toString();
+							req.setAttribute("AssignedList", service.ActionSelfList(EmpId));
+						}catch(Exception e){
 								e.printStackTrace();
 								logger.error(new Date() +" Inside ActionSelfList.htm "+UserId, e);
 						}
-						
-
 						return "action/ActionSelf";
 					}
